Change how core snap mounts are created
This patch fundamentally changes how core snap bind mounts are created
when running on a classic distribution. To align with how an all-snap
distribution looks like, the root filesystem is now a read only core
snap. From that snap a number of bind mounts reach to the classic system
(for user and system writable data and a few other things).

Lastly since the runtime environment is now identical in all places, we
should reset PATH to always aligned with the core snap.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>