Implement the rootfs-is-core-snap config switch
This patch implements the build-time option to use the core snap as the root
file system. This arrangement differs from the older approach is that the host
directory structure is more indepented as all the bind mounts are made from a
specific host directory to a well-defined directory in the core snap.

It also is closer in principle to how an all-snap system operates, where the
initrd mounts the core snap as rootfs and then bind mounts additional things
from the writable area.

In the future the host side can become driven by snapd so that any host-specifc
logic can be used. This also opens the path for using interfaces to create
dedicated bind mounts (e.g. to share fonts).

Signed-off-by: Zygmunt Krynicki <zkrynicki@gmail.com>