Tweak VENDOR_ARGS
This patch changes VENDOR_ARGS in two ways, first on Ubuntu we get
—enable-nvidia-ubuntu. In the past this was always on which was never
really tested

Second, now —enable-roofs-is-core-snap is enabled by default. All the
distributions (including Ubuntu) are shipping it now so it’s time for
Debian to follow.

Lastly (did I say two?) the —disable-confinement option is replaced by
—disable-apparmor.