Allow running qemu spread tests offline
This patch allows spread tests to work offline. It requires some
explicit setup that is described in spread-tests/offline-content/README

The main motivator for this change is that it makes the testing cycle
work while on the plane/bus/train or where connectivity is limited or
unreliable. Even with reliable conncetion it can cut test time
significantly. On my relatively slow machine a full spread run for
xenial can now run in roughly 9 minutes. As a comparison, the same tests
ran on linode for nearly 11 minutes. I don't have hard numbers for
online tests on my machine (offline also makes some expensive operations
cached) but the initial time was closer to 20 minutes.

As a secondary factor testing is now a little bit more reproducible and
repatable. Spread makes ordering non-deterministic but at least package
building and snap revisions are fixed.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>