Fix creation of $XDG_RUNTIME_DIR and $SNAP_USER_DATA
This patch fixes two^Hthree issues in the creation of the aforementioned
directories.

The first fix is to create those directories with elevated permissions.
This allows snap-confine to create subdirectories of /run, namely
/run/user, /run/user/1234 and finally /run/user/1234/snap.$SNAP_NAME.
Without this fix when snap-confine runs as a non-root user (ignoring
setuid root bit for a second) it cannot create sub-directories in
/run/user/.

None of the current tests captured this because they use root natively
(via spread). A spread test in snapd used su to switch to a test user
and that reliably failed without this patch.

The second fix is to change the owner of those directories to what the
real_uid and real_gid is. This makes both directories usable by the
intended user.

The final fix was a mistake in mkpath(), if mkpath ever encountered a
directory (e.g. /) it would carry the EEXIST error all the way out.
This would prevent the ownership changing code from doing its job. This
is fixed by resetting errno before each mkdirat(2) call.

In addition to code changes the apparmor profile has been changed to
allow access to @{HOME}/snap/*/*/ and /run/user/*/*/. This may be
redundant but is the correct way to go given that we use mkpath() to
create those directories.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>