Allow sc_nonfatal_mkpath() to control each segment
This patch changes the aforementioned function to have a callback that
controls the mode and ownership of each created segment. The callback is
optional, if omitted the mode and ownership of each created segment is
the same.

All exiting callers are modified to pass the new options structure that
conveys all of this information. This automatically fixes directories
controlled through $SNAP_USER_DATA and $XDG_RUNTIME_DIR environment
variables. They are now tied with helper functions that return the
correct data for each segment.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>